Split vc-api and test-app into workspaces
This extracts the vc issuer's API into a new workspace and changes the
test-app from a standalone package to a workspace of the main package.

By extracting the issuer API, the types (like credential request) can be
reused elsewhere.

By turning the test-app into a workspace, we deduplicate a lot of the
dependency information (by removing on package lock) and we can use
local dependencies (like the issuer types).
