+++ src/vc-api/README.md
@@ -1,0 +1,11 @@
+# VC Issuer API
+
+A small TypeScript library with type & [zod](https://zod.dev) definitions for the [issuer API](../../demos/vc_issuer/README.md).
+
+## Usage
+
+Build from the root directory:
+
+```bash
+$ npm run --workspace ./src/vc-api build
+```
