+++ src/vc-api/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./../../tsconfig.json",
+  "include": ["src"],
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "composite": true
+  }
+}
