+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["src"],
+  "include": ["src/showcase", "src/frontend"],
   "compilerOptions": {
     "allowJs": true,
     "checkJs": false,
